requirements/base.txt
asn1crypto==0.22.0
cffi==1.10.0
cryptography==1.8.1
Django==1.10.5
django-allauth==0.31.0
django-appconf==1.0.2
django-compressor==2.1.1
django-grappelli==2.9.1
django-webpack-loader==0.4.1
djangorestframework==3.6.2
enum34==1.1.6
idna==2.5
ipaddress==1.0.18
oauthlib==2.0.2
packaging==16.8
paramiko==2.1.2
pyasn1==0.2.3
pycparser==2.17
pyparsing==2.2.0
python-openid==2.2.5
rcssmin==1.0.6
requests==2.13.0
requests-oauthlib==0.8.0
rjsmin==1.0.12
six==1.10.0
